requirements.txt
audioop-lts==0.2.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cryptography==45.0.6
filelock==3.18.0
Flask==3.1.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.7.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pillow==11.3.0
pycparser==2.22
pypiwin32==223
python-dotenv==1.1.1
pyttsx3==2.99
pywin32==311
regex==2025.7.34
requests==2.32.4
scipy==1.16.1
semantic-version==2.10.0
setuptools==80.9.0
setuptools-rust==1.11.1
sounddevice==0.5.2
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tiktoken==0.10.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
flask
flask-mysqldb
python-dotenv
cryptography
gunicorn
